requirements.txt
aiohttp
python-decouple
https://github.com/New-dev0/Telethon/archive/Aurium.zip
html_telegraph_poster
